refactor: Extract parsing logic from Linter
refs #18787
